vim: change hexman for vinarise
Vinarise is still in development but IMHO is better than Hexman.
The main reason is that you can save the file in Hexadecimal mode without end
with a piece of shit! :smile:
Another reasons are that I think that is a very well thought and is more
powerful. And don't mix the ASCII and Hexadecimal editors, if you want to change
an ASCII value, edit it in Vim.