vim: delete the Sessions plugin.
When we save a session with multiple buffers & tabs, it no will load correctly
the session saved. The behaviour is erratic, only works well (and not so much)
when we use the default Vim commands for the task (mksession/source). And the
plugin's author is not willing to maintain it because is not currently using it
and thinks that the Vim's default session behaviour is strange (after multiple
tests, I think that too).