vim: store the neobundle log in a file
Now the NeoBundle log includes the date (relative time) and revision number when
a plugin is updated. Is something like that:

```
[neobundle/install] (63/75): |syntastic| Updated
[neobundle/install] |syntastic| * 2d9ff24 [19 minutes ago] Minor cleanup: remove
dead code; formatting.
[neobundle/install] Installed/Updated bundles:
syntastic
```

This can be a great help to find out when a plugin was updated in order to
solve a situation in which a plugin is not working properly due to a error
introduced in a previous revision. Like I said in the Vim README file:

> Since we often install plugins from repositories, we are exposed to error-
prone updates. Once in a while, a plugin update introduces a bug and you end up
with an unstable configuration until that bug is fixed (which may take a while).
If we use Vim for work, that is very inconvenient.

This new feature can help us to restore manually the plugin to a previous state
when still worked fine (thanks to the relative time date). We can do this by
deleting the plugin and reinstalling it again using the `revision lock` feature
of NeoBundle in the `.vimrc` file.