vim: remove the vim-multiple-cursors plugin.
I decided to remove this plugin because is sloppy, slowly with many selections
and have multiple annoying bugs and conflicts with another plugins. And by the
way, I almost never use it, is not very useful.

It can be still being used with this configuration, simply adding the line

    NeoBundle 'joedicastro/vim-multiple-cursors'

to the ~/.vim/custom.vim file.