vim: now the auto-selection of the optimal word in auto-complete pop-up dialog is disabled by default, but can be enabled/disabled again via a unite entry or a mapping.
The reason to do this is that in too much common situations, when you want to
write very fast, the auto-complete auto-selection feature can be a serious
obstacle to the productivity, mainly in editing/writing simple text. However
is very easy to enable it again with a mapping and toggle it again at one's
discretion.