emacs: remove the init.el file
I used to include the init.el file in the repository, but now I removed
it and added an explanation of how to generate it from the readme.org
file. This make easier to me the management of the changes of the emacs
configuration in this repository, without to have to duplicate the sames
changes in both files (mostly when I'm doing partial commits).